requirements.txt
streamlit==1.40.1
nltk==3.9.1
spacy==3.7.5
scikit-learn==1.5.2
gensim==4.3.3
pandas==2.2.3
wikipedia==1.4.0
sumy==0.11.0
pdfplumber==0.11.9
python-docx==1.1.2
matplotlib==3.10.8
